Temp fixed module cache issue

+++ MaceSwinger/src/com/maceswinger/utils/CustomMouse.java
@@ -1,21 +1,21 @@
 package com.maceswinger.utils;
 
 import org.lwjgl.input.Mouse;
-
-import com.maceswinger.client.GameClient;
+import org.lwjgl.util.vector.Vector2f;
 
 public class CustomMouse {
 	
 	
 	
-	public static double getX(){
+	public static float getX(){
 		return Mouse.getX()*CustomDisplay.getxScale();
 		
 	}
-	public static double getY(){
+	public static float getY(){
 		return Mouse.getY()*CustomDisplay.getyScale();
 		
 	}
+	
 	public static boolean isButtonDown(int i) {
 		return Mouse.isButtonDown(i);
 	}
